# Conflicts:
#	Physics/Source/Scene01.cpp

+++ Physics/Source/Player.cpp
@@ -55,7 +55,7 @@ void Player::Update(double dt)
 		//Physics::K1(playerBomb->vel, Vector3(0.f, -9.8f, 0.f), (float)dt, playerBomb->vel);
 		//playerBomb->pos += playerBomb->vel * (float)dt * 40.f;
 		if (launched && !exploded)
-			bombspin -= 50*dt;
+			bombspin -= 50*(float)dt;
 		if (playerBomb->pos.y < 0)
 		{
 			playerBomb->SetActive(false);
@@ -63,7 +63,7 @@ void Player::Update(double dt)
 		}
 		else if (exploded && playerBomb->scale.x < 10)
 		{
-			playerBomb->scale = Vector3(playerBomb->scale.x + 10 * dt, playerBomb->scale.y + 10 * dt,1);
+			playerBomb->scale = Vector3(playerBomb->scale.x + 10 * (float)dt, playerBomb->scale.y + 10 * (float)dt,1);
 			if (playerBomb->scale.x > 10)
 			{
 				playerBomb->SetActive(false);
@@ -175,16 +175,16 @@ void Player::Upgrade(Tree::avl_node &node)
 //		playerObj->scale.Set(10, 10, 1);
 //		playerObj->mass = 1.5f;
 		playerObj->scale.Set(playerObj->scale.x + (node.item_count * 1), playerObj->scale.y + (node.item_count * 1), 1);
-		playerObj->mass = 1 + (node.item_count * 0.1);
+		playerObj->mass = 1 + (node.item_count * 0.1f);
 		//m_speed = 20;
 	}
 	else if (node.id == 2)
 	{
-		m_speed = 10 + (node.item_count * 1);
+		m_speed = (float)(10 + (node.item_count * 1));
 	}
 	else if (node.id == 3)
 	{
-		jump_boost = (node.item_count * 30);
+		jump_boost = (float)(node.item_count * 30);
 	}
 	else if (node.id == 4)
 	{
